requirements.txt
flask
flask-swagger-ui
pytesseract
pillow
flask-cors